Write tests for UserManager

+++ app/src/main/java/comp3350/intellicards/Business/UserManager.java
@@ -11,6 +11,10 @@ public class UserManager {
         userPersistence = Services.getUserPersistence();
     }
 
+    public UserManager(UserPersistence persistence) {
+        userPersistence = persistence;
+    }
+
     public boolean registerUser(String username, String password) {
         if (userPersistence.getUserByUsername(username) != null || username.equals("guest")) {
             return false; // Username already exists or it equals 'guest' which is reserved for guest mode
@@ -28,7 +32,11 @@ public class UserManager {
         return null;
     }
 
-    public void deleteUser(String username) {
-        userPersistence.deleteUser(username);
+    public boolean deleteUser(String username) {
+        if (userPersistence.getUserByUsername(username) != null && !username.equals("guest")) {
+            userPersistence.deleteUser(username);
+            return true;
+        }
+        return false;
     }
 }
